Clean up send form

+++ lib/data/sources/repositories/compose_repository_impl.dart
@@ -11,7 +11,7 @@ class ComposeRepositoryImpl extends ComposeRepository {
 
   @override
   Future<RawTransaction> composeSend(
-      String sourceAddress, String destination, String asset, double quantity,
+      String sourceAddress, String destination, String asset, int quantity,
       [bool? allowUnconfirmedTx, int? fee]) async {
     final response = await api.composeSend(
         sourceAddress, destination, asset, quantity, allowUnconfirmedTx, fee);
